На проверку второй вариант
